% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_cld.R
\name{ask_gulati}
\alias{ask_gulati}
\title{Solution recommendng engine based upon the sayings of gulati sir}
\usage{
ask_gulati()
}
\value{
null
}
\description{
This program is based on the views of gulati sir used as an experiment for academic purposes only
}
